+++ Shape_detection/include/CGAL/Shape_detection/Efficient_RANSAC/Shape_base.h
@@ -501,7 +501,8 @@ namespace CGAL {
       int score = 0;
       std::vector<std::size_t> indices(num);
       for (std::size_t i = 0;i<num;i++)
-        indices[i] = m_indices[get_default_random()(m_indices.size())];
+        indices[i] = m_indices[get_default_random()(
+          static_cast<unsigned int>(m_indices.size()))];
       
       std::vector<FT> dists(num), angles(num);
       other->squared_distance(indices, dists);
@@ -515,7 +516,8 @@ namespace CGAL {
         return false;
 
       for (std::size_t i = 0;i<num;i++)
-        indices[i] = other->m_indices[get_default_random()(other->m_indices.size())];
+        indices[i] = other->m_indices[get_default_random()(
+          static_cast<unsigned int>(other->m_indices.size()))];
 
       this->squared_distance(indices, dists);
       this->cos_to_normal(indices, angles);
